refactor: remove ovh prefix
