--- src/ng-tail-logs.html
@@ -1,4 +1,0 @@
-<div class="ovh-angular-tail-logs" data-scroll-glue>
-    <div data-ng-transclude></div>
-    <i class="ovh-angular-tail-logs__cursor" />
-</div>
