+++ src/template.html
@@ -1,0 +1,4 @@
+<div class="tail-logs" data-scroll-glue>
+    <div data-ng-transclude></div>
+    <i class="tail-logs__cursor" />
+</div>
